# docs/swagger.yaml
swagger: "2.0"
info:
  title: Move Big Rocks API
  description: |
    Self-hosted operations software for support workflows, agent access, and installable extensions.

    ## Authentication
    Move Big Rocks uses magic link authentication. Request a magic link via email, then use the session cookie for authenticated requests.

    ## Multi-Tenant Architecture
    All workspace-scoped endpoints require authentication and workspace context. Admin endpoints require instance operator privileges.

    ## Optional Extensions
    Move Big Rocks core provides shared primitives such as workspaces, cases, contacts, forms, attachments, automation, and agent access.
    Optional first-party extensions may add Sentry-compatible error tracking, web analytics, and other specialized product surfaces.

    ## API Surface
    The stable REST contract for webhooks/forms runs on `api.movebigrocks.com`.
    Internal admin workflow routes below are on `admin.movebigrocks.com/admin/actions`; these are non-contractual admin UI internals.
    Admin GraphQL helpers use `POST /admin/graphql` (session-authenticated admin pages), while `api.movebigrocks.com` uses `/graphql` for agent and CLI traffic.
    `api.movebigrocks.com` serves `/graphql` for agent and CLI traffic.
  version: "1.0.0"
  contact:
    name: API Support
    email: support@movebigrocks.com
  license:
    name: Proprietary
    url: https://movebigrocks.com/license
  termsOfService: https://movebigrocks.com/terms
host: api.movebigrocks.com
basePath: /
schemes:
...
securityDefinitions:
  SessionCookie:
    type: apiKey
    in: cookie
    name: mbr_session
    description: Session cookie obtained after magic link verification
  BearerAuth:
    type: apiKey
    in: header
    name: Authorization
    description: 'Bearer token format: "Bearer {jwt_token}"'
...
paths:
...
  /v1/webhooks/postmark/{secret}/inbound:
    post:
      tags:
        - Webhooks
      summary: Postmark inbound email webhook
      description: Receives inbound emails from Postmark and creates support cases
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: secret
          type: string
          required: true
          description: Webhook secret token
        - in: body
          name: body
          required: true
          schema:
            type: object
            description: Postmark inbound email payload
      responses:
        200:
          description: Email accepted
          schema:
            type: object
            properties:
              status:
                type: string
                example: accepted
              email_id:
                type: string
        401:
          description: Invalid webhook secret
          schema:
            $ref: '#/definitions/ErrorResponse'
  /v1/webhooks/postmark/{secret}/bounce:
    post:
      tags:
        - Webhooks
      summary: Postmark bounce webhook
      description: Receives bounce notifications from Postmark
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: secret
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            description: Postmark bounce payload
      responses:
        200:
          description: Bounce processed
          schema:
            type: object
            properties:
              status:
                type: string
              email_id:
                type: string
        401:
          description: Invalid webhook secret
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  ErrorResponse:
    type: object
    properties:
      error:
        type: object
        properties:
          type:
            type: string
          message:
            type: string
          code:
            type: string
          details:
            type: object
      timestamp:
        type: string
        format: date-time
      request_id:
        type: string
      path:
        type: string
...